+++ flecsi/exec/mpi/task_prologue.hh
@@ -4,16 +4,11 @@
 #ifndef FLECSI_EXEC_MPI_TASK_PROLOGUE_HH
 #define FLECSI_EXEC_MPI_TASK_PROLOGUE_HH
 
-#include <flecsi-config.h>
-
+#include "flecsi/config.hh"
 #include "flecsi/data/privilege.hh"
 #include "flecsi/data/topology.hh"
 #include "flecsi/exec/mpi/future.hh"
 #include "flecsi/util/demangle.hh"
-
-#if !defined(FLECSI_ENABLE_MPI)
-#error FLECSI_ENABLE_MPI not defined! This file depends on MPI!
-#endif
 
 #include <mpi.h>