+++ flecsi/run/options.hh
@@ -236,6 +236,8 @@ private:
 
 }; // struct program_option
 
+/// \}
+
 } // namespace flecsi
 
 #endif