+++ flecsi/util/bitutils.hh
@@ -20,7 +20,7 @@ constexpr T
 bit_width(T x) noexcept {
   static_assert(std::is_unsigned_v<T>);
   T ret = 0, d = std::numeric_limits<T>::digits;
-  const auto high = [&] { return x >= T(1) << d - 1; };
+  const auto high = [&] { return x >= T(1) << (d - 1); };
   if(high())
     return d; // avoid overwide >>= below
   // Perform a binary search for the first 1.  The static number of iterations