+++ flecsi/util/common.hh
@@ -4,7 +4,7 @@
 #ifndef FLECSI_UTIL_COMMON_HH
 #define FLECSI_UTIL_COMMON_HH
 
-#include <flecsi-config.h>
+#include "flecsi/config.hh"
 
 #include <algorithm>
 #include <cassert>